Cleanup ContainerFactory files
